ASV_analysis

This function performs the analysis on ASV sequences identified by the previous steps and aligns them to the reference sequence. First, it pools together those sequences characterized by a Hamming distance equals or lower than 2, summing their counts. To perform this step, EvoTraceR uses the sequence clusering agorithm impemented in the python package UMI-tools Then it performs pairwise alignment using Needleman-Wunsch global alignment algorithm implemented in function pairwiseAlignment in package pwalign, aligning each sequence to the original barcode considered in the analysis. (See the pwalign documentation here for more details). After identifying all indels (insertions are identified by their start position and number of nucleoides inserted, while deletions are identified by their start and end position), it removes the ones that are too small and don't span any cut site: to perform this filter, it exapnds the start and end position by the number of bases specified in the parameter cleaning_window, it counts the number of cut sites spanned after the expantion and it removes those that span 0 sites. Next, it pools together those sequences that differ by one another only by substitutions. Then it computes the normalized counts of each ASV in each sample, dividing each count by the total number of sequences for each sample and multiplying by 1e6, yelding Counts Per Million (CPM). After normalization, counts are summed in each ASV, and those sequences showing a frequency lower than the threshold specified in parameter asv_count_cutoff are removed. Finally, it performs filtering of the sequences based on their flanking sequences.

Usage

asv_analysis(
  EvoTraceR_object,
  ref_name = "BC10v0",
  ref_seq =
    "TCTACACGCGCGTTCAACCGAGGAAAACTACACACACGTTCAACCACGGTTTTTTACACACGCATTCAACCACGGACTGCTACACACGCACTCAACCGTGGATATTTACATACTCGTTCAACCGTGGATTGTTACACCCGCGTTCAACCAGGGTCAGATACACCCACGTTCAACCGTGGTACTATACTCGGGCATTCAACCGCGGCTTTCTGCACACGCCTACAACCGCGGAACTATACACGTGCATTCACCCGTGGATC",
  ref_flank_left = "^TCTAC",
  ref_flank_right = "CCCGTGGATC$",
  flanking_filtering = "right",
  ref_cut_sites = c(17, 43, 69, 95, 121, 147, 173, 199, 225, 251),
  ref_border_sites = c(1, 26, 52, 78, 104, 130, 156, 182, 208, 234),
  output_figures = TRUE,
  asv_count_cutoff = 2,
  pwa_gapOpening = -25,
  pwa_gapExtension = 0,
  pwa_match = 15,
  pwa_mismatch = -4,
  pwa_type = "global",
  cleaning_window = c(3, 3),
  batch_size = 100,
  cores = parallel::detectCores()
)

Arguments

EvoTraceR_object

(Required). Object of class EvoTraceR, result of the function initialize_EvoTraceR

ref_name

String indicating the ID of the reference sequence used in the experiment. Default is 'BC10v0',

ref_seq

String indicating the reference sequence used in the experimenti. Default is 'TCTACACGCGCGTTCAACCGAGGAAAACTACACACACGTTCAACCACGGTTTTTTACACACGCATTCAACCACGGACTGCTACACACGCACTCAACCGTGGATATTTACATACTCGTTCAACCGTGGATTGTTACACCCGCGTTCAACCAGGGTCAGATACACCCACGTTCAACCGTGGTACTATACTCGGGCATTCAACCGCGGCTTTCTGCACACGCCTACAACCGCGGAACTATACACGTGCATTCACCCGTGGATC',

ref_flank_left

String indicating the first nucleotides of the reference sequence that never mutate over the course of the experiment. Default is "^TCTAC",

ref_flank_right

String indicating the first nucleotides of the reference sequence that never mutate over the course of the experiment. Default is "CCCGTGGATC$",

flanking_filtering

Which among the flaning regions to use to filter out contaminated sequences. Must be one of c('left', 'right', 'both', 'either'), default is 'both'

ref_cut_sites

Positions in the reference sequence of the cutting sites. Default is c(17, 42, 68, 94, 120, 146, 171, 198, 224, 251),

ref_border_sites

c(26, 52, 78, 104, 130, 156, 182, 208, 234).

output_figures

(Optional). Default TRUE: Boolean indicating whether a user whishes to store a figure indicating the number of ASV tracked during the different steps of the analysis.

asv_count_cutoff

(Optional). Default to 2. Minimum number of counts in Counts Per Million (CPM) for an ASV to be considered in the statistics.

pwa_gapOpening

(Optional). Default is -25. Parameter gapOpening passed to pairwiseAlignment from pwalign (See description).

pwa_gapExtension

(Optional). Default is 0. Parameter gapExtension passed to pairwiseAlignment from pwalign (See description). Default is 0.

pwa_match

(Optional). Default is 15. Parameter indicating the score for matches during pairwise alignment with pwalign. This parameter, together with the previous one, are used to construct the substitution matrix used by the function pairwiseAlignment.

pwa_mismatch

(Optional). Default is -4. Parameter indicating the penalty for mismatch events during pairwise alignment with pwalign.

pwa_type

(Optional). Parameter indicating the type of pairwise alignment. Must be one of One of "global", "local", "overlap", "global-local", and "local-global". For more details see original documentation

cleaning_window

(Optional). Default is c(3,3). Vector containing the number of nucleotides that we use for extending respectively the start and end position of each indel to determine the ones that don't span any cut sites and thus get removed. (See description for more information).

batch_size

(Optional). Default is 100. Number of batches to split reads into for parallel execution of pairwiseAlignment. This parameter can be tuned together with the cores parameter to optimize the speed of alignment.

cores

(Optional). Default is parallel::detectCores(). Number of cores to use for pairwise alignment.

Value

The EvoTraceR object passed as a parameter with the following new fields:

clean_asv_dataframe: ASV sequences identified post-filtering (contamination removed, sequences with a similarity higher than pid_cutoff_nmbc to the original barcode aggregated to it and ASVs named in increasing order (ASV01, ASV02, etc.) according to their total counts.

reference: Info about the reference sequence used for the current analysis.

statistics: another list with the following sub-fields:

asv_df_percentages: dataframe with six columns. asv_names is the name of the ASV. sample is the sample identifier (e.g. ID of an organ or, in case of longitudinal data, of the timepoint); count: total counts per million for a specific ASV in a specific sample; perc_in_sample: CPM normalized to the total counts in the corresponding sample; perc_asv: CPM normalized to the total counts for the corresponding ASV; perc_fold_to_max: CPM normalized to the maximum counts observed for the corresponding ASV in a sample.

asv_totalCounts: for each ASV, total counts and number of samples in which it was detected.

sample_totalcounts: for each sample, total counts and number of distinct ASVs detected.

asv_diversity_persample: measures of clonal richness and measures of heterogeneity computed for each sample based on the ASVs detected.

asv_persample_frequency: counts for each ASV in each sample.

asv_persample_detection: binary matrix indicating whether a sequence has been detected in the corresponding sample.

asv_toBarcode_similarity: edit distance, percentage similarity and alignment score of each ASV compared to the original barcode.

all_asv_statistics: all the statistics computed on each ASV grouped together in the same tibble.

alignment, another list with the following fields:

Binary_mutation_matrix: binary matrix encoding the presence/absence of a mutation in an ASV.

asv_barcode_alignment: tibble where each line corresponds to a position in a ASV, and the columns encode the following information:

asv_names: name of the ASV

sample: sample identifier

position_bc260: position of the alteration in the original barcode. Note that insertions are assigned to the position that coincides with their beginning.

alt: type of alteration. wt = Wild Type (i.e. non-mutated position). sub = substitution. del = deletion. ins = insertion.

ref_asv,

read_asv: respectively, the reference nucleotide observed in the original barcode and the one observed on the sequence.

mutations_coordinates: dataframe containing a list of all mutations, with their start and end position.

ASV_alterations_width: dataframe containing the number of nucleotides affected by each type of mutation in each ASV.

This function saves the figure asv_filtering_freq.pdf, which is a barplot indicating the change in the number of sequences during all the processing steps.

Details

Then, using the CPM it computes different statistics for the final ASVS, storing: the relative frequency of all ASVs in each sample. the relative frequency of each ASV in the samples. the counts for each ASV normalized to the counts of the sample with maximum frequency. the frequency of the different ASVs in each sample.

Examples

if (FALSE) { # \dontrun{
data(EvoTraceR_object)
EvoTraceR_object = asv_analysis(EvoTraceR_object = EvoTraceR_object)
} # }