infer_phylogeny

This function reconstructs the phylogenetic tree using the Cassiopeia suite. It uses the greedy algorithm, that iteratively splits the set of ASVs in clusters based on the most common mutation at each iteration. Note that this requires installing cassiopeia in the anaconda/virtual enviroment used by reticulate.

Usage

infer_phylogeny(EvoTraceR_object, mutations_use = "del_ins")

Arguments

EvoTraceR_object

(Required)

mutations_use

(optional). Default = 'del_ins' A string indicating what type of mutations to use for phylogeny reconstruction. Can be one of c('del', 'del_ins'). The first corresponds to using only cleaned deletions. The second refers to the case where cleaned deletions and insertions are considered.

Value

EvoTraceR object with a new field called phylogeny, that stores:

mutations_in_phylogeny: string indicating which mutations were used for phylogeny reconstruction.

tree: dataframe storing the tree returned b y Cassiopeia. This dataframe has a column "group" that indicates the clonal population to which the node belongs to (a clonal population is a cluster returned by Cassiopeia).

tree_phylo: phylo object containing the tree returned by Cassiopeia.

tree_uncollapsed: phylo object containing the tree returned by Cassiopeia prior to the collapsing of mutationless edges.

Examples

if (FALSE) { # \dontrun{
EvoTraceR_object = infer_phylogeny(EvoTraceR_object)
} # }